refactor: rename components
Images -> ImageList
ImageItem -> ListItem
