+++ src/style/Images/ListItemStyle.ts
